Update test.py
When using cuda, there will be a RuntimeError (Expected object of backend CPU but got backend CUDA for sequence element 1 in sequence argument at position #1 'tensors' ). After debugging it, i found that it was because class_vector was not tensor type. So i change this line and it can run well.